backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==3.2.3
django-debug-toolbar==3.2.3
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
exceptiongroup==1.1.3
flake8==6.1.0
flake8-isort==6.0.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.3
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.5
webcolors==1.11.1
gunicorn==20.1.0
